Run one of rstan, cmdstanr or cmdstan on a file of Stan code

Run one of rstan, cmdstanr or cmdstan on a file of Stan code

run_stan_interfaces(
  input_to_stan,
  path_to_stan_code,
  interface = c("rstan", "cmdstanr", "cmdstan"),
  iter_warmup = 250,
  iter_sampling = 250,
  chains = 4,
  cores = parallel::detectCores(),
  params_to_ignore = c("lp__", "rho[1,1]", "rho[2,2]", "rho[3,3]", "rho[4,4]",
    "rho[2,1]", "rho[3,1]", "rho[4,1]", "rho[3,2]", "rho[4,2]", "rho[4,3]", "rho.1.1",
    "rho.2.2", "rho.3.3", "rho.4.4", "rho.2.1", "rho.3.1", "rho.4.1", "rho.3.2",
    "rho.4.2", "rho.4.3", "f_plate_effects_unscaled",
    "f_effects_by_pred_var_cat_unscaled", "p_pos_effects_by_pred_var_cat_unscaled",
    "mu_pos_effects_by_pred_var_cat_unscaled", "sd_pos_effects_by_pred_var_cat_unscaled",
    "mu_neg_effects_by_pred_var_cat_unscaled", "sd_neg_effects_by_pred_var_cat_unscaled",
    
     "y_cal_mean_per_obs", "y_cal_mean_per_obs", "y_sam_mean_per_obs",
    "y_obs_sd_cal", "y_obs_sd_sam", "exp_f_1_mult_f_4_per_plate", "p_pos_log_per_sam_id",
    "p_neg_log_per_sam_id", "p_pos_per_sam_id", "mu_pos_per_sam_id", "mu_neg_per_sam_id",
    "sd_pos_per_sam_id", "sd_neg_per_sam_id", "f_3_min_f_2_per_plate", "p_blank_log",
    "p_blank_log1m", "y_obs_sd_min_log_shift_unscaled",
    "y_obs_sd_jump_log_shift_unscaled", "y_obs_sd_sam_min_per_plate",
    "y_obs_sd_cal_min_per_plate", "y_obs_sd_sam_jump_per_plate", 
    
    "y_obs_sd_cal_jump_per_plate"),
  downsampling_factor = 1L,
  cmdstan_path_to_installation = NA,
  cmdstan_path_to_json = NA,
  cmdstan_overwrite_json = FALSE,
  cmdstan_path_to_output = NA,
  cmdstan_path_to_compiled_model = stringr::str_remove(path_to_stan_code, ".stan$"),
  cmdstan_read_output_into_df = TRUE,
  ...
)

Arguments

input_to_stan

a list containing all the input the Stan code expects.

path_to_stan_code

the path to the file containing the Stan code.

interface

one of "rstan", "cmdstanr" or "cmdstan".

iter_warmup

a positive integer: the number of warmup iterations per chain (during which the sampling algorithm adapts; these are excluded from the output).

iter_sampling

a positive integer: the number of sampling iterations per chain (which are included in the output).

chains

a positive integer: the number of chains used for sampling.

cores

a positive integer: the number cores used in parallel for computation.

params_to_ignore

a character vector naming parameters to be excluded from output (if possible; interface dependent).

downsampling_factor

a positive integer: the factor by which to downsample the posterior. e.g. if a value of 2 is specified, we keep 1 in every 2 samples. The default of 1 means we keep all samples. TODO: currently only implemented for cmdstan.

cmdstan_path_to_installation

the path to where cmdstan is installed on your system; you need to specify this if interface="cmdstan", but not otherwise. Inside this directory there should be an executable file named make (which we use to compile Stan code).

cmdstan_path_to_json

the path to where we will write a temporary json file to hold the input for cmdstan; you need to specify this if interface="cmdstan", but not otherwise.

cmdstan_overwrite_json

a single logical value: should we overwrite a file at cmdstan_path_to_json if it exists already?

cmdstan_path_to_output

the path to where we will write output files from cmdstan; you need to specify this if interface="cmdstan", but not otherwise. Several files will be created with things appended to this path: _chain1.csv, _chain2.csv etc.

cmdstan_path_to_compiled_model

the path where we will create the compiled version of the stan code. Some value (such as the default) is needed if interface="cmdstan", but not otherwise.

cmdstan_read_output_into_df

a single logical value: should we read cmdstan output files into a dataframe that is returned by this function? If a value FALSE is specified, this function returns a value NULL.

...

additional arguments will be passed to rstan::sampling() (if interface="rstan") or to the $sample() method of the cmdstanr::CmdStanModel() object (if interface="cmdstanr").

Value

a dataframe with one row per sample from the posterior and one column per parameter (unless interface is set to cmdstan and cmdstan_read_output_into_df is set to FALSE).